DNS/requirements.txt
dnslib
dnspython